requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.9
albumentations>=1.3
opencv-python>=4.7
Pillow>=9.0
scikit-learn>=1.2
pandas>=1.5
numpy>=1.23
matplotlib>=3.7
seaborn>=0.12
omegaconf>=2.3
grad-cam>=1.4
kaggle
jupyter
notebook
ipykernel
torchinfo
pytest>=7.0
pytest-cov>=4.0
ruff>=0.1
